Split big main file into smaller more manageable files
